@@ -11,14 +11,16 @@
     ],
     "require": {
         "php": ">=8.1",
-        "symfony/framework-bundle": "^6.0|^7.0",
-        "symfony/console": "^6.0|^7.0",
-        "symfony/process": "^6.0|^7.0",
-        "symfony/event-dispatcher": "^6.0|^7.0",
-        "symfony/http-kernel": "^6.0|^7.0",
-        "symfony/http-foundation": "^6.0|^7.0",
-        "symfony/dependency-injection": "^6.0|^7.0",
         "symfony/config": "^6.0|^7.0",
+        "symfony/console": "^6.0|^7.0",
+        "symfony/dependency-injection": "^6.0|^7.0",
+        "symfony/event-dispatcher": "^6.0|^7.0",
+        "symfony/framework-bundle": "^6.0|^7.0",
+        "symfony/http-foundation": "^6.0|^7.0",
+        "symfony/http-kernel": "^6.0|^7.0",
+        "symfony/messenger": "^7.3",
+        "symfony/process": "^6.0|^7.0",
+        "symfony/scheduler": "^7.3",
         "symfony/yaml": "^6.0|^7.0"
     },
     "require-dev": {
